requirements.txt
a2a-sdk[all]==0.3.22
fastapi
uvicorn
redis
pydantic>=2.0.0
gradio>=4.0.0
pyyaml
openai
httpx
